fix: code refactor, optimization
